fix tipo da variavel preco na class produto

+++ models/produtos/Produto.java
@@ -4,7 +4,7 @@ public abstract class Produto {
 
     private String nome;
 
-    private float preco;
+    private Double preco;
 
     private int quantidade;
